Conflicts:
	src/main/java/dnd/jackpot/project/dto/ProjectSearchDto.java

+++ src/main/java/dnd/jackpot/project/dto/CommentDto.java
@@ -11,5 +11,6 @@ public class CommentDto {
 	
 	private boolean privacy;
 	
-	private int projectId;
+	private long projectId;
+	
 }
